requirements.txt
pillow
pyyaml